pdMBM: add 8-bpp RLE support
Pretty straightforward!  I'm not breaking my back to perf-test this aggressively, given the typical image sizes of this format, but the code is nicely performant as-is.  Results across Symbian's test-suite appear to be uniformly accurate (yay).

One caveat is that I'm not entirely sure how RLE plays with DWORD-alignment, since RLE runs and chunks can extend across scanline boundaries.  Ignoring alignment is currently producing good images for me, but that may change upon testing 16- and 24-bpp RLE (which have more complex alignment rules as-is).  If so, I'll revisit this code.