Lots of plumbing for run-time window resizing and animation effects
PhotoDemon has a lot of backend code for providing real-time previews of still-image effects.  It does *not* have a similar mechanism for previewing animation effects.  It's time to change that!

This commit lays much of the foundation for run-time resizing of windows with underlying animation previews.  Working from the new-ish Image > Animation window, I'm slowly adding the code necessary for windows like this to be user-resizable, while hopefully not explding PCs due to the intense resource requirements of live animation previews (where each frame needs some kind of effect applied).