Finish plumbing for run-time animation preview resizing
This new solution works very well, and it effectively makes animation previews as efficient as still-image previews during window resize operations.  (In fact, it's so good that I now have to kill off my old custom solution in the Image > Animation tool, so I can replace it with this!)

This was the major blocker for starting work on actual effects in the new Image > Animation menu, so that's my task for today!

(Thank you also to Roy K for an updated German localization in this commit)