Start laying groundwork for 3D LUT export
This feature will (hopefully) allow you to export any combination of PD color corrections to a 3D LUT file (only .3dl and .cube planned right now).  That file can then be used to apply those color corrections using any other photo editor.

The biggest challenge here will be generating an actual LUT matrix.  I want to automate the process as much as possible, so PD needs to (somehow?) build a reference image in the background for you, then silently apply all of your current actions to that reference image in order to produce a matrix that covers the full color spectrum.  (Most other software requires you to *start* with a reference image, or they limit you to a specific subset of actions a la Photoshop.  I want to be more flexible than this.)

The complexity of this level of automation is why I'm breaking this job into more stages than usual.  No current ETA on when this feature will be ready.