MBM parser: add support for masks in AIF files
Relates to #341

It's hard to know how robust my code is when I only have so many AIF files to test, but this version of the parser correctly retrieves (and applies) masks on 100+ images randomly sampled from downloads available at this page:

https://web.archive.org/web/20170803124306/http://www.pscience5.net/colouricons.htm

That's good enough for me.  If anyone encounters AIF files that do *not* appear to parse correctly, please send them to me so I can investigate.

I've also added *.abw and *.acl extensions to PD's common-dialog interface.  Thank you to @Kroc for the tip!