PSP decoder: layers can now be imported
There's a *lot* of clean-up and testing work left to do (as you can see from the hundreds of "TODO" comments scattered throughout the code), but the core fundamentals are now in place.  Multi-layer PSP files can now be opened by PhotoDemon, and PD will import as many layers as it is currently capable of understanding.

I've only tested semi-modern PSP images (not old v5 ones which are a different beast) and I've primarily tested normal RGBA color-depths.  More aggressive testing of esoteric color-depths and sizes is forthcoming.