PSP decoder: add support for layer groups
Note that this is for raster groups only; vector groups are a different beast in PSP files (with a totally different encoding).

As with PSD files, PhotoDemon will now generate "dummy" layers that mark the start and end of each group inside the PSP.  It will also apply group visibility to child layers, so that if a layer is visible but it sits within an invisible group, PD will mark it as *not* visible.

This is obviously an unideal workaround until PD supports real layer groups, but the beauty of it is that group import code is already done!  The only thing that will change in the future is tying the import code to PD's "real" layer group mechanism, when such a thing arrives.

Also, PSP's group implementation is garbage and I hate it.  Instead of placing simple group start/end markers in the file (like Photoshop does), they only embed group *end* markers.  Each group end marker also stores a count of how many layers are in that group - but obnoxiously, the count omits grandchildren, great-grandchildren etc, so you have to manually maintain a record of tree-depth and count each layer as you encounter it to try and solve where each group begins.  It's overly complex and silly and it's especially problematic for 3rd-party decoders like mine, because we can't use every layer in the source image but I still have to track unuseable layers so I know where to place group start markers uuuuuuuugh

Anyway, this is precisely why I like to write file format decoders before adding certain features to PD.  It's very helpful to study how other software behaves, so I don't fall into the same mistakes when I add their features to PhotoDemon.