PSP decoder: fix rendering of vector layers with active selections or masks
This trove of JASC-produced sample images (that shipped with PSP 6) is proving to a be a huge boon for testing.  Some of the files in the collection are vector-based, but these vector layers have empty masks and/or selection channels attached for reasons unknown.  (I say unknown because the layers and masks don't show up in PSP itself... so I suppose it could have just been a bug in their encoder which was patched out in later versions?  idk)

Regardless, PD was detecting these pixel channels and overwriting vector data with the pixel-based selection and/or mask data instead, which we obviously don't want.  Because PD's vector renderer is working well, the decoder will now defer to it instead of embedded pixel data when it encounters vector layers.