ExifTool: update to latest version
My ExifTool interface is a coding disaster (entirely of my own making), so I always fear diving into it.  I had to tweak a few things to make it play nicely with recent ExifTool changes, and I *think* everything is working again.  One of these days I'll migrate to a proper XML parsing solution, but alas today is not that day...