Effects > Artistic > Colored Pencil: revert to original version
Late last year, I started experimenting with an alternate implementation of this effect.  I'd love to pursue that line of exploration, but I've got too many other features on my plate right now, and reworking esoteric effects is not the best use of my time.

This at least ensures that nightly builds ship with a useable version of the effect.