AVIF support: offer to automatically download libavif if it's missing
This is ultimately the best solution, I think.

Because the AVIF support libraries are both massive (4x bigger than PhotoDemon itself!) and only available in 64-bit, PD will not ship with them.  But on compatible systems, PD will still allow the user to attempt to load and/or save AVIF images.

If the user tries to load or save an AVIF image and libavif is missing, PhotoDemon now displays a descriptive message and offers to download the ~20mb libraries for you.  (This only needs to be done once, and the user can opt-out.)

If the user agrees, PD will automatically download and configure the latest tested libavif release (0.9.0 at the time of this writing).  If the user was attempting to load an AVIF file, the import will automatically proceed like any other normal import.

If the user opts-out of the automatic download, PD gives them some information on how they can manually enable AVIF support.  The active import, if any, is canceled.

Unfortunately, this commit includes a *lot* of new translation strings.  My apologies to translators!