PSD decoder: harden against incompatible PSD files
This solves a rare edge-case when loading HDR PSD images (16+ bits per channel) comprised solely of vector or fill layers.  Photoshop always includes raster copies of these layers in SDR images, but *not* in HDR images (presumably to keep file-size down).  Because PD can't currently interpret PSD vector layers, and there is no raster backup, it would end up producing a 0-layer image that would then cause the program to crash when it tried to access the image.

I've now reworked the PSD loader so that it performs a final failsafe check against the pdImage object produced by the parser.  If the file was marked as valid but no usable layers were recovered, the PSD parser will now reset itself and resume parsing at the merged image offset (if one exists).  This results in a single-layer copy of the merged PSD being loaded, which is much more useful than a broken 0-layer "null" image.