New framework for advanced image resize filters
PD has historically used either built-in Windows libraries (GDI+) or the 3rd-party FreeImage library for image resizing.  This provided nice coverage of standard image resampling algorithms, with good performance to boot.

But sometimes we want even *better* quality than what a 3rd-party library can provide.  That's the goal with this branch: to finally provide a comprehensive, native-to-PhotoDemon image resampling library.

My work here is heavily based on an original image resampling library shared by Libor Tinka:

https://www.codeproject.com/Articles/11143/Image-Resizing-outperform-GDI

Many thanks to Libor for sharing his work under a liberal license.

This first draft is a very rough translation of the original C# source, with myriad changes for VB compatibility and/or performance quirks.  I expect many more changes will be necessary to get this adequately working in PD, but at least this is a start.