Image resampler: progress bar now works
Progress bar display during resize hasn't historically worked, because 3rd-party libraries (like GDI+) were handling the resize for us and they don't report progress.  But PD's new internal resamplers fix this!

(Note that the behavior is slightly different when resizing an image with many layers - in this case, the resampler instead uses the number of layers completed as the progress bar tracker, instead of reporting progress on just the *active* layer)