Image > Resize: first draft of "live previews" feature
Previews for this tool have to be designed (and coded) from scratch because they bear no similarity to other effects and adjustments.  That makes this challenge considerable, which is why no other photo editors (aside from the latest editions of Photoshop, I suppose) have attempted it.

I've got this first pass working OK, but there's a lot left to do.  Performance is the biggest challenge, since downsizing a huge image requires processing of pretty much the full source image in order to produce a useful preview.  I have some ideas for improving performance, but they're all complicated and not-guaranteed-to-work... but hey, at least this first step is pretty cool!