Resampling engine: fix serious perf issue on new integer-only version
This was a real surprise to me, but basically mistyping a single multiplication factor from an integer value to a floating-point value resulted in a 10+X performance hit on my little Ryzen-2 laptop (150 ms to 1.7 seconds).  Sheesh.  I hope I haven't made mistakes like that elsewhere in the project...

Anyway, the cool thing about fixing this is that the new integer-only resampler is 2x faster than the old floating-point one, with no noticeable quality difference!  I'm committing this little fix now because I'm about to attempt a further rework for even better performance.  If *that* works, I'll extend these fixes to the y-direction as well for another small boost.