Image Resampler: new integer-based version ready for primetime!
PD now provides two internal resampling engines for resizing an image: a "perfect accuracy" floating-point-based one, and a "nearly perfect accuracy" integer-based one.  The integer-based one produces results that are 99.8 to 99.9% identical (via PSNR) and it has 2x or better performance on modern CPUs, with even larger gains on older CPUs.  There's really no reason (outside of theoretical curiosity) to not use the integer version, but PD allows users to toggle this using the already existing "allow optimizations" checkbox on the resize dialog.  (Note that for the "live preview", the integer-based transform will always be used due to performance considerations.)

I've got one last potential optimization to try and improve live preview performance, but it will be preview-mode-specific.  The resampling engine itself is pretty much finished and ready for prime-time (barring a bit of static analysis to remove unused variables).