Image > Resize: now supports click-drag to pan around the preview image
As mentioned in previous commits, the preview box in this dialog needs to be completely custom-built, so little features like this have to be reimplemented from scratch (ugh).

Also in this commit are some UI clean-ups, a refactor of the way previews interact with the "fast" mode toggle, and... I'm officially dropping FreeImage's resampling engine entirely.  It's barely faster than PD's custom-built engine, and it requires more memory (since FreeImage allocates its own buffers, and we have to copy our data into/out of those buffers).  These trade-offs no longer make sense given the quality and performance of PD's new pure-VB resampling engine, so I'm deactivating FreeImage's resize interop in production builds.  (The curious can still re-enable this via a constant at the top of FormResize, if they want.)

The resize dialog has also been made resizable, and the preview area automatically "grows" to match the area of the resized window.

With this, I think the new Resize tool is ready for merging!  I'm very excited to get wider testing on this feature, as it's a rare accomplishment to have live previews of resampling in a free photo editor.  (To my knowledge, only Photoshop provides previews of image resize operations - and even then, they only added it in Photoshop CC.  I am not aware of any free photo editors that provide something similar... until now, hehe)