New Layer > Replace menu
Relates to #317.  Thank you to @jpbro for this suggestion, and my apologies for taking so long to implement it.

PhotoDemon now provides a Layer > Replace menu which has three options for updating the current layer's contents in-place:

1) From the clipboard
2) From an arbitrary image file
3) From the current image (visible layers merged together)

These options provide a convenient way to update the contents of an existing layer from an outside source, without the hassle of jumping through various Copy+Paste operations.