fix potential error in new Effects > Distort >Droste
If a border of the effect lines up with a border of the image, our use of Single-precision values can produce small negative numbers due to rounding issues around zero in the formula.  This causes a subsequent Log() calculation to crash.  Easy fix is to Abs() potential problem values before we Log 'em.