Fix _MouseUp() potentially getting missed if a tooltip is active
Fixes #365.  Many thanks to @mizka76 for catching and reporting!

This was an incredibly obnoxious bug to track down, and it may be limited to Win 10 (or some specific subset of Windows versions including Win 10).  If a tooltip is active on a given control, and the tooltip is animating a fade-in or fade-out when the underlying control is clicked, and the tooltip window is within several pixels of the mouse event, the subsequent _MouseUp event could go missing, causing PhotoDemon to *think* the mouse was still over the target window even though it was not!  This would cause the next _MouseDown event to still get sent to the old button - which is an issue with Windows, not PD.

This is not the first time I've encountered oddities with input behavior while a window is animating, and thankfully PD only relies on built-in OS window animations for tooltip fade-in/fade-out.  Similarly, this was really only problematic on the main window Undo/Redo buttons, whose tooltips fade in to show the name of the command being undone/redone.  The timing of the tooltip's appearance and clicking those buttons has a much higher probability of overlap than elsewhere in the program.  I'm just glad mizka76 caught the association between this bug and those buttons, which made it possible to track this incredibly pesky issue down.