Edit > Paste: fix unnecessary error message when pasting from certain webbrowsers
If PD finds HTML on the clipboard that looks like an image link, it'll try to download it (which spares us from the headache of whatever a given webbrowser has done to an image internally).  Unfortunately, these snippets are sometimes not accurate and PD has to settle for the browser's copy of the image.  When this happens, we don't need to raise an error message about the initial download failure.