WebP export: dialog now uses libwebp for export preview
This is a placeholder commit as I finally got core export code working (ugh it's complicated).  But previewing an exported WebP file works when using libwebp as the preview encoder/decoder, and there are no memory leaks which is especially exciting.

Now to clean-up the code, fully implement all export options, and add libwebp code for the *actual* save process (not just the preview).