Fix potential bug in MergeVisibleLayers
When merging two layers, a third (temporary) surface needs to be created to hold the result of the merge.  This is because the two layers may not be an identical size and/or may not overlap precisely, so the size of the merged layer may be larger than either layer individually.

PhotoDemon internally skipped the need for that third surface if a caller explicitly promised that the base layer was the size of the image.  (This is precisely what happens when doing MergeVisibleLayers - a new blank layer is created at the size of the image, and all visible layers are merged onto that.)  This lets us skip the need for a temporary merge surface, and greatly reduces memory churn if merging a huge amount of layers one-after-another.

Unfortunately, I hadn't considered the case where the top layer in the merge is larger than the *entire* image itself!  (Which is a valid condition, technically.)  When this happens, we still need to create a temporary merge surface, since a bottom layer the size of the canvas *still* isn't large enough to hold the merge results.

Anyway, I've completely done away with the parameter that lets the caller "guarantee" a certain surface size, and instead the merge function validates layer sizes manually.  If certain conditions are met, it will still reuse the base surface as-is, so the optimization still exists - it's just properly turned off if the top layer is bigger than the canvas itself.