pdAccelerator: refacor in light of new hotkey design
I honestly had trouble remembering how this user control worked (or why it was designed the way it is).  A small refactor with a focus on code reduction was very helpful, and I've improved comment quality to better explain some of this control's quirks.  Some unnecessary trackers and parameters have also been removed, putting this control in a better place to fulfill its new role: *just* tracking keypresses and leaving the actual hotkey management to the new Hotkey module