New Actions engine for executing high-level commands
Menus, hotkeys, and the search bar all need to execute specific actions in response to user input.  Each of these input mechanisms previously handled their own (lengthy) list of command translations.

I have now condensed all of these into a single "Actions" module.  Just call Actions.LaunchAction_ByName() with the name of the action you want to perform, and it handles the rest - including not launching invalid operations (e.g. ask it for Image > Resize, and it will ignore you if no images are open).

This was a very large change touching many parts of the program.  As usual, I couldn't help myself and I fixed a number of other problems while I was here.  Sorry/not sorry.

- Removed unnecessary global variables by converting them to internal values accessible by safe wrapper functions.
- Reworked old 3rd-party library interactions (FreeImage especially) with modern PD classes, for much less code and reduced chance of errors.
- Demoted the 3rd-party FreeImage library to "low-priority" status.  This means it is loaded much later in the program initialization chain, which improves perceived startup time on old PCs.
- Rewrote the few remaining menus that rely on menu captions for triggering actions.  These now use internal action IDs, and the old "launch action from caption" code has been removed.
- Added new action triggers for some operations that were previously only accessible by menu.
- Reorganized the main dialog's code to be roughly alphabetically ordered, and standardized all comments and code layout for menus.
- Removed old main dialog UI code that was no longer necessary
- Improved validation for some action targets

This is a huge codebase improvement that has been some time in the making.  It's also possible I broke some actions while here - I'll be testing this more comprehensively in the coming days.