Fix window parent+child relationship on right toolbox
The right-side toolbox uses a more complex nesting relationship (to enable vertically resizeable panels), and window bits were not being correctly set to ensure that FormMain remained the top parent in the chain.  This would make FormMain's titlebar look inactive when a panel on the right toolbox had focus.

All is well now, though it requires some messy window bit juggling to keep it this way.