Fix main menu keyboard navigation
Uuuugh, this was a frustrating one.  I'm not sure if this is a recent Win 10 break or if I just missed it earlier, but when an Alt key is used to access the menu (e.g. Alt+F for File), the window that previously had focus does NOT receive a "Lost focus" notification (via either WM_ACTIVATE with flags or WM_KILLFOCUS).  That window will keep eating keypresses because it has no idea it lost focus to the menu.

Working around this requires querying main menu state (and system menu state, because we're thorough) before handling keypresses inside a keyboard hook.