Large memory reductions when working on multiple images simultaneously
When loading a new image, PD will now automatically suspend inactive images to compressed buffers.  This is a very fast operation (lz4 is used) and for "normal" image sizes - e.g. < 20 megapixels - any delays caused by the suspend process are basically unnoticeable.  On larger images there may be a brief stutter when switching rapidly between loaded images, but the given the huge memory reductions, I think this is a worthwhile modification.

In the future it may be nice to have an option to go even further and suspend-to-disk if memory is constrained, but for now, compressed in-memory buffers provide a fast, simple way to cut memory usage when working with multiple side-by-side images.  (I may look at extending this auto-suspend feature to other places, for example prior to initializing demanding tools.)