Image > Animation tool: support window Maximize button
I don't know if I want to enable maximize button support on all tool windows - because real-time previews simply do not perform well on many tools when the preview area is that large - but I can start with supporting the animation window since it's an exception to the rule.

Anyway, the window maximize button provides a quick way to get a full-screen preview of a large animation.  (Manually resizing a window to that size is unwieldy and annoying.)