Macro recording: specifically disallow "Original image" state
Relates to #377 .  Thank you to @TyraVex for catching and reporting.

(To be completely honest, I'm not sure why that action was being recorded in the first place - but PhotoDemon will now specifically check for it during both recording and playback, and ignore it if found.  This ensures no one needs to re-record buggy macros.)

While I was here, I continued my eternal quest to simplify and condense localized text.  Updated localization files forthcoming.