Pivot in flyout design
To work correctly (without a ton of hacks), these need to be children of the main window.  This limits their display area to the main window's boundaries, but honestly, I think this works better.  (In fact, I want to go a step further and limit it to specifically the canvas area (so toolboxes remain unobstructed).)

This is basically how flyout panels work in Photoshop.  I'm still double-checking to make sure the panel is auto-closed correctly under all circumstances (particularly when losing focus, which is complicated as a single panel may host myriad controls), but so far this is looking workable.  Long and tedious, but workable.