Another attempt at fixing "mouse getting stuck on a control" behavior
As part of overhauling PD's toolboxes (WIP in another branch), I've started testing the app on an old XP VM deliberately stripped down to single-core and very low RAM limitations.  That made the ongoing "mouse gets stuck on a random control" bug skyrocket in reproducibility, which is both good and bad (good because I have a new way to test fixes, bad because it's an intensely frustrating bug).

This simple change appears to fix it in early testing.  I did not put useful comments in the commit where I originally modified this line of code, so I have no idea if this revert carries repercussions I should beware of... but because this provides an immediate fix for a vexing issue, I'm going to revert it while I keep testing additional possibilities.