Move/size toolbox: redesign complete
With this, I think the viability of the new flyout design is proven, and it's time to start migrating other toolpanels to this new layout.  I'm excited to see how it works, although I do anticipate some complexities with panels like the text tools.

Oh well, one problem at a time!