Measure tool: convert to new flyout design
Relates to #333

This entire panel has been reworked with a new dynamic run-time layout system.  I like it!

With this, the first set of tools (non-destructive group) is officially migrated to the new toolbar system.  Very exciting.

Next up... paint tools, maybe?  I could tackle text tools, but whew... those stress me out.  I honestly don't know where to begin with them because the panels are *so* dense.

Selection tools are also an option, but I'm actually going to merge the in-progress "selection tool overhaul" branch into this one first, as it includes a number of new UI settings for selections (like combine modes) and I don't want to merge those after-the-fact because it will be incredibly ugly.

So lots left to do, but this massive job is trucking along nicely.