Flyout panels: new buttons to "lock" panels open
Part of #333

PD's new toolbox flyout panels automatically close when they lose focus.  In 99% of cases this is preferred behavior, but just in case a user needs a panel to stay open, flyout panels now provide a "lock" button in the bottom corner.  A "locked" panel will stay open even if the panel loses focus.  (Tooltips explain this in more detail.)

This is theoretically helpful if a panel provides stuff that a user wants to interact with *while* interacting with the canvas (or menu-driven actions).

Locked panels will still close if the current tool is deactivated, or if a new panel is opened.  (Panels may overlap when dropped, and for my own sanity I need to limit them to one-open-at-a-time.)