Initial batch of work on selection toolbox flyouts
Relates to #333

Had to switch the icon for locking flyout panels (to an eye instead of a lock) because locks are already used elsewhere on the selection panels, and it's incredibly confusing to reuse them for panel locking.  I probably need to make a new icon with an eye and lock together, sigh.

Anyway, selection tools are complicated!  All selection tools share the same options panel, but have different subpanels for type-specific settings.  This makes the new  flyout design more complicated than previous tools.  I also want to ensure that useful information is not hidden during interactions, so I'm adding code for auto-displaying meaningful flyouts when e.g. moving or resizing a selection.  This all adds complexity.

Anyway, I've just tackled rectangular selections so far but flyouts are working well and I don't foresee any "deal-breaking" complications, so I suppose this is going as well as could be expected