Fix tab-key tracking when switching toolboxes
Ugh, I should have caught this sooner.

Previously, PD would completely unload a toolpanel when the user switched away from it.  This was a way to keep memory usage at a minimum.

But that also incurred an unpleasant delay when switching between tools, so I've since abandoned that and just cache panels persistently in memory.  Unfortunately, PD's tab-key tracker (which determines ideal tab order at run-time and does not rely on tab-index values) was not updated correctly to account for this new behavior.

Now it has been.