New toolpanel engine: add basic flyout animation
Relates to #333

AnimateWindow is a terrible API, but I don't have time to write my own replacement.  So I'm going to use AnimateWindow for awhile to see if basic flyout animations are an improvement or not.  It's pretty janky on my local PCs but it's possible that is due to all the WIP on the flyout engine.

If this ultimately proves obnoxious, it's a simple constant flip to deactivate these animations later.