Clone brush: expose advanced brush features...
...and also, migrate to the new flyout UI!

(Relates to #333)

Advanced brush features like dab spacing have always been available to this tool (because it shares PD's central brush engine), but there was not enough room on the old toolbar to show them.  With the new flyout design, however, I can cram in as many settings as I want!  Yay...?

(The long-term plan is a new unified brush selector with supports for saved presets, brush shapes, etc, a la Photoshop - but one problem at a time.)

With this, there are only two tools left to convert to the new UI...