Gradient tool: migrate to new flyout UI
Relates to #333

This is the final tool.  The bulk of this horrifying project is done!

I imagine there will still be plenty of tweaking required to get layouts as identical as possible across different tools, and I may need to make different decisions about how settings are grouped in certain tools (to maximize discovery), but at least the new UI framework is in place and all on-canvas tools *work* again.

Next up: finalizing text so I can generate new localization files for nightly builds