Pencil tool: prototype for auto-moving flyouts into the top panel...
...when there's enough horizontal space to do so.

I like how nicely this works, but wow - it's a *lot* of code.  And this is just for handling one setting.  (Ultimately I'd like to move "as many settings as possible" into the parent panel if there's room for them, but I no longer know how feasible that is.)

I may be able to generalize a solution to this, but it's tricky because which settings get moved into the parent panel really depend on each tool - ideally the most commonly used settings are always available in the toolpanel itself, with more esoteric settings confined to flyouts.

I'm gonna need to think this over for a bit before adding this feature to other tools, to make sure I'm going about it the right way.  (I'm increasingly  worried about a maintenance nightmare if I've got hundreds of lines of code on each tool for handling setting migration between flyouts and parent panels, especially if I find bugs down the road.)