Selection tools: auto-hide dropped panels *if* the mouse intersects them
Relates to #382

Selection tools are tricky because we want to know their precise position (if moving them) or size (if resizing/creating them).  There is not enough room to display both on the selection toolpanel, so PhotoDemon tries to be useful by automatically dropping the relevant flyout so you can see these measurements while interacting with a selection.

A downside with this is that the dropped flyout panel may obscure your view of the canvas.  To remedy this, PhotoDemon will now constantly verify the current mouse position against the flyout rectangle.  If the two overlap, the flyout will automatically be hidden.

(Still todo: a toggle on the flyout to prevent auto-display behavior entirely.)