All selection tools: display boundary rect in status bar
Relates to #382

When editing with any selection tool, PD's status bar will now display the boundary rectangle dimensions of the current selection.  This matches the current UI convention in other photo editors.

(The current icon used in the status bar is just a placeholder; I'll update it shortly but I need to do it from a different PC!)