Add icon for new "selection boundary size in status bar" feature
Relates to #382

PhotoDemon's status bar will now display a persistent notification of the active selection's boundary dimensions.  I could also display more information here if useful, but I'm not sure that e.g. the total area or top-left coordinate position is relevant when switching away from the active selection tool.  (That's the main purpose of displaying selection size in the status bar - so you can reference it even when you *aren't* using a selection tool.)

This change brings PD in line with similar behavior in Photoshop, GIMP, and Paint.NET, at least.

Thank you to @g-busch and @Kroc for their assistance with this ongoing selection-tool-refinement work