First-time "theme and language" dialog: add new explanation...
... that theme and language settings can be changed at any time from the Tools Menu.

I've also added a note that additional customizations are available in the Window menu, since users may not think to check there.  (The Window menu is where Photoshop, GIMP et al stick their workspace settings, so I think it  makes sense to stick with this design for now... but it is a little jarring to force users to both the Tools and Window menus for UI customizations.  Alas I don't have a better organization at present *shrug* )