Proof-of-concept for new "move selected pixels" tool
Relates to #303

My current idea is to simply implement this feature as part of the normal move/size tool (instead of creating a dedicated tool for it, like Paint.NET).  This is how Photoshop works.

As implemented in this commit, Undo/Redo doesn't work and the clipboard is used for interop so this is obviously not a viable permanent solution.  More this just demonstrates that this approach *should* work once a variety of kinks are worked out.

I will likely need to add new options to the Move/Size tool to make this more useful (e.g. copying from the active layer or all layers, etc) - and in fact, these additions would allow for some behaviors that are not easily achieved even in Photoshop.  Worth a try!

I haven't fully decided if this tool should be implemented before or after my current multiple-selections-at-once work, but this at least gives a nice starting point.