Fix Undo/Redo behavior on export dialogs
Metadata (on export dialogs) is handled by a dedicated control.  This control generates a random filename for holding temporary metadata - as required by ExifTool.

The control was generating this temporary filename on-demand, but now that dialogs support their own mini Undo/Redo behavior, the control was being queryed on a timer system and generating a bunch of random filenames.  This is obviously unnecessary - just generate a random filename once, at initialization, then hand out that same filename on any query.